Initial commit of Celebrity Facematch Ai Demo
The demo detects a face from a camera live stream and runs the face
through a neural network to obtain face detail embeddings. This step can
be utilized on the NPU of the i.MX 8MPlus.

The embeddings are afterwards compared to the top 1000 celebrities from
IMBD to select the celebrity which looks most alike the detected face.

The demo is written in python and utilizes openCV, tflite and a GUI
based on GTK+ 3.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>