ai: Reduce initial warmup time
As stated in [1] the initial warmup time of the NPU can be significantly
reduced by OVX graph caching which caches the associated graph of a
model in an .nb file. It is activated by two envirnment variables.

In our demo we were able to reduce the warmup time of the NPU from 55 to
9 seconds.

[1] https://www.nxp.com/docs/en/application-note/AN12964.pdf

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>