Activation of the NPU accelerator
Enabled the NPU accelerator by adding the path to the external delegate library
which exists inside the model initialisation function. This allows utilizing
the hardware accelerator driver (OpenVX) to use the accelerator capabilities.
The NPU had to be enabled inside the ai.py code since it is not the default
settings in the hardknott release of the NXP.

Signed-off-by: Mohamed Aly <m.aly@phytec.de>