ai.py: Changing the preprocessing function for the mobilenet_v2 model
Changed the preprocessing function to the one compatible with the mobilenet_v2
trained model. The image pixels are scaled betweeen 0 and 1 by dividing the
input samples by 255.0 in float format. Also, removed the input parameters
(data_format, version) from the preprocessing function which are not needed
here. The data format will be picked automatically from the keras_backend stored
in the /.keras/keras.json file.

Signed-off-by: Mohamed Aly <m.aly@phytec.de>